+++ Trabalho_java/src/main/java/com/Integrado/Trabalho_java/repository/CursoRepository.java
@@ -3,5 +3,5 @@ package com.Integrado.Trabalho_java.repository;
 import com.Integrado.Trabalho_java.model.Curso;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CursoRepository extends JpaRepository<Curso, Integer> {
+public interface CursoRepository extends JpaRepository<Curso, Long> {
 }
