+++ Trabalho_java/src/main/java/com/Integrado/Trabalho_java/repository/ProfessorRepository.java
@@ -3,5 +3,5 @@ package com.Integrado.Trabalho_java.repository;
 import com.Integrado.Trabalho_java.model.Professor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProfessorRepository extends JpaRepository<Professor, Integer> {
+public interface ProfessorRepository extends JpaRepository<Professor, Long> {
 }
