+++ Trabalho_java/src/main/java/com/Integrado/Trabalho_java/repository/TurmaRepository.java
@@ -3,5 +3,5 @@ package com.Integrado.Trabalho_java.repository;
 import com.Integrado.Trabalho_java.model.Turma;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TurmaRepository extends JpaRepository<Turma, Integer> {
+public interface TurmaRepository extends JpaRepository<Turma, Long> {
 }
